Use int indices and simplify channel sorting
Replace size_t loop/index types with int throughout channelizer.c to ensure consistent signed indexing and avoid signed/unsigned issues. Update printf format for channel prints to %d. Simplify and switch get_sorted_channel_indices to use int (update declaration in channelizer.h) and replace the previous even/odd special-case logic with a single contiguous index ordering that groups higher-frequency bins followed by lower ones. These changes harmonize index types across the file and update the header signature to match.

+++ channelizer.h
@@ -24,7 +24,7 @@ typedef struct {
     double split_filter[NUM_CHANNELS][COEF_PER_STAGE];
 } channelizer_handle;
 
-void get_sorted_channel_indices(size_t num_channels, size_t *sorted_idx);
+void get_sorted_channel_indices(int num_channels, int *sorted_idx);
 int channelizer_setup(channelizer_handle *handle);
 void *channelizer_thread(void *arg);
